Remove dead code related to building single vs multi arch on Mac
Since we only build for one architecture on Mac now, this code is not needed.
Plus it would only run if we were on Leopard and we don't build on Leopard anymore.

Also, never add "--no-multiarch" to the check command on Mac even if
STAGE4_mode != "multiarch" as that flag has no effect.